Changes xp types to be int instead of float, fixes issues

+++ priv/migrations/001_create-sessions.up.sql
@@ -1,7 +1,7 @@
 CREATE TABLE IF NOT EXISTS sessions (
   id INTEGER PRIMARY KEY AUTOINCREMENT,
-  required_xp REAL,
-  xp REAL,
+  required_xp INTEGER,
+  xp INTEGER,
   is_active BOOLEAN DEFAULT true,
   created_at TEXT,
   updated_at TEXT
